refactor: move DuplicateStep
